Name the end-of-game specs after the file they test
chess-ducks keeps its moves in helpers.ts, so its legality and end-of-game
tests join up in helpers.spec.ts.
