+++ src/components/games/chess-bishops/chess-bishops.spec.ts
