+++ src/components/games/chess-knight/chess-knight.spec.ts
