+++ src/components/games/chess-rook/chess-rook.spec.ts
