+++ src/components/games/dominoes-on-chessboard/dominoes-on-chessboard.spec.ts
