+++ src/components/games/rook-to-corner/rook-to-corner.spec.ts
