+++ src/components/games/triangle-coloring/triangle-coloring.spec.ts
